+++ fxn/api/__init__.py
@@ -5,9 +5,9 @@
 
 from .dtype import Dtype
 from .environment import EnvironmentVariable
-from .feature import Feature
 from .prediction import CloudPrediction, EdgePrediction, Prediction
 from .predictor import Acceleration, AccessMode, EnumerationMember, Parameter, Predictor, PredictorStatus, PredictorType, Signature
 from .profile import Profile
 from .storage import Storage, UploadType
 from .user import User
+from .value import Value